feat: overall upgrade to PHP 7.4 compatibility

+++ app/services/uploadHandler.php
@@ -314,7 +314,7 @@ class UploadHandler {
      * @param string $str
      */
     protected function toBytes($str){
-        $val = trim($str);
+        $val = (int) trim($str);
         $last = strtolower($str[strlen($str)-1]);
         switch($last) {
             case 'g': $val *= 1024;
